Reporting, Corpse visuals, Game class refactor
Added report button when near a corpse. Doesn't do anything yet.

Corpses are more like the players that died

Refactored many game concepts into a shared Game class.

When game state changes, the change is put into netvar classes and JSON'd and broadcast to players in the game, who then turn it into a local version of the Game class, and update the game appropriately.

Some refactoring of USE for simplicity. More to go.

No fadeoverlay in practice.

Details:
There is now an actual practice game.

With regards to data layout, I'm not crazy about the technical possibility of bugs from players existing in multiple games.