Meeting player portraits
Meetings:
Visible player portrait in meetings
Players sorted by playernum and living.
Added DEV_STARTMEETING to test meetings.

Details:
sh_gamestate:
Moved any "on changed to this gamestate" type logic to GameStateChanged(), it was in multiple places before and sometimes was skipped.

Server and client run a Shared_OnGameStateChanged_PerPlayer whenever a player's character is added.

cl_fadeoverlay:
Replaced playernum2visible with a function and used it in all places where this happens.

sh_utils:
Added ClonePlayerModel() and some color utility functions.
Replaced all Players.LocalPlayer with GetLocalPlayer() which asserts that we're on the client.