Easier Put Books Away, higher doorways
Increased doorway height so tall characters are not impacted.

Put books away:
Increased margin for error, and an extra 4x on touch devices.

When points are rewarded the gain is now drawn over the matchscreen, so when a player is voted off it is more obvious that you get points for it.

details:
cl_coins:Copy the UI into 3 different UIs on startup, for each of the three main areas of focus, so they can be ordered independently.