No playernums in other rooms, checkers back to Great Room
Checkers is back in the Great Room but not in the lobby.

No longer draw PlayerNums for players in other rooms.

Added some environmental details to some upstairs rooms.

details:

sh_rooms:
Added not_in_lobby bool that makes a task not appear in the lobby.

cl_rooms:
Track which room players are in on the client, instead of just local player.
Put all players into START_ROOM by default, so able to get rid of GetCurrentRoom and EMPTY_ROOM.