Lobby Cycling
If the lobby is out of date, send players to a limbo server so the lobby server will die.

Details:

Track server version in persistence. If the server discovers that it no longer matches, then send all the players to a reserved server with a flag to get sent back to lobby after they get there.

sv_matchmaking:
Refactor matchmaking thread into several functions for clarity.